refactoring; Updated README.md file;
