+++ src/test/java/pu/jpa/dynamicquery/model/projection/NameProjection.java
@@ -1,5 +1,7 @@
 package pu.jpa.dynamicquery.model.projection;
 
+import pu.jpa.dynamicquery.api.Projection;
+
 /**
  * @author Plamen Uzunov
  */
