+++ src/test/java/pu/jpa/dynamicquery/model/projection/ProductProjection.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import pu.jpa.dynamicquery.api.Projection;
 
 /**
  * @author Plamen Uzunov
